#336: Moved the PassiveCounterAttackParser to the model folder
 The file was not changed yet.
