+++ src/app/ffbe/model/effects/passives/passive-counter-attack.parser.ts
@@ -1,5 +1,5 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
+import {EffectParser} from '../../../mappers/effects/effect-parser';
+import {Skill} from '../../skill.model';
 
 export class PassiveCounterAttackParser extends EffectParser {
   public parse(effect: Array<any>, skill: Skill): string {
